refactor autopilot
