+++ android/app/src/main/java/org/openbot/autopilot/AutopilotFragment.java
@@ -89,8 +89,7 @@ public class AutopilotFragment extends CameraFragment {
     binding.cameraToggle.setOnClickListener(v -> toggleCamera());
 
     List<String> models =
-        getModelNames(
-            f -> f.type.equals(Model.TYPE.AUTOPILOT) && f.pathType != Model.PATH_TYPE.URL);
+        getModelNames(f -> f.type.equals(Model.TYPE.CMDNAV) && f.pathType != Model.PATH_TYPE.URL);
     initModelSpinner(binding.modelSpinner, models, preferencesManager.getAutopilotModel());
     initServerSpinner(binding.serverSpinner);
 
@@ -219,7 +218,7 @@ public class AutopilotFragment extends CameraFragment {
     try {
       Timber.d(
           "Creating autopilot (model=%s, device=%s, numThreads=%d)", model, device, numThreads);
-      autopilot = Autopilot.create(requireActivity(), model, device, numThreads);
+      autopilot = new Autopilot(requireActivity(), model, device, numThreads);
       croppedBitmap =
           Bitmap.createBitmap(
               autopilot.getImageSizeX(), autopilot.getImageSizeY(), Bitmap.Config.ARGB_8888);
@@ -237,7 +236,10 @@ public class AutopilotFragment extends CameraFragment {
               () ->
                   binding.inputResolution.setText(
                       String.format(
-                          "%dx%d", autopilot.getImageSizeX(), autopilot.getImageSizeY())));
+                          Locale.getDefault(),
+                          "%dx%d",
+                          autopilot.getImageSizeX(),
+                          autopilot.getImageSizeY())));
 
       Matrix cropToFrameTransform = new Matrix();
       frameToCropTransform.invert(cropToFrameTransform);
